Refactor: 사용자 데이터 가져오기 로직 커스텀 훅으로 분리
